:truck: Moved common Zipper definition to new file GenericZipper
